pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ralph-footy"
version = "0.1.0"
description = "NRL Footy Forecaster — tip well, teach well."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "James Roberts" },
]

dependencies = [
    "anthropic",
    "python-dotenv",
    "requests",
    "rich",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
ralph = "ralph.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["ralph"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
